% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_change_password}
\alias{lemmy_change_password}
\title{Change your user password.}
\usage{
lemmy_change_password(
  new_password,
  new_password_verify,
  old_password,
  auth,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{new_password}{A character value.}

\item{new_password_verify}{A character value.}

\item{old_password}{A character value.}

\item{auth}{A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 3:
\preformatted{<list>
|-jwt: (Optional) A character value
|-registration_created: A logical value
`-verify_email_sent: A logical value
}
}
\description{
\code{HTTP.PUT /user/change_password}
}
