% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_delete_account}
\alias{lemmy_delete_account}
\title{Delete your account.}
\usage{
lemmy_delete_account(
  password,
  auth,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{password}{A character value.}

\item{auth}{A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 0:
\preformatted{<list>
}
}
\description{
\code{HTTP.POST /user/delete_account}
}
