% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_delete_custom_emoji}
\alias{lemmy_delete_custom_emoji}
\title{Delete a custom emoji}
\usage{
lemmy_delete_custom_emoji(
  id,
  auth,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{id}{A numeric value.}

\item{auth}{A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 2:

\if{html}{\out{<div class="sourceCode">}}\preformatted{<list>
├─id: A numeric value
└─success: A logical value
}\if{html}{\out{</div>}}
}
\description{
\code{HTTP.Post /custom_emoji/delete}
}
