% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_get_post}
\alias{lemmy_get_post}
\title{Get / fetch a post.}
\usage{
lemmy_get_post(
  id = NULL,
  comment_id = NULL,
  auth = NULL,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{id}{(Optional) A numeric value.}

\item{comment_id}{(Optional) A numeric value.}

\item{auth}{(Optional) A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 4:
\preformatted{<list>
|-post_view: <list>
| |-post: <list>
| | |-id: A numeric value
| | |-name: A character value
| | |-url: (Optional) A character value
| | |-body: (Optional) A character value
| | |-creator_id: A numeric value
| | |-community_id: A numeric value
| | |-removed: A logical value
| | |-locked: A logical value
| | |-published: A character value
| | |-updated: (Optional) A character value
| | |-deleted: A logical value
| | |-nsfw: A logical value
| | |-embed_title: (Optional) A character value
| | |-embed_description: (Optional) A character value
| | |-thumbnail_url: (Optional) A character value
| | |-ap_id: A character value
| | |-local: A logical value
| | |-embed_video_url: (Optional) A character value
| | |-language_id: A numeric value
| | |-featured_community: A logical value
| | `-featured_local: A logical value
| |-creator: <list>
| | |-id: A numeric value
| | |-name: A character value
| | |-display_name: (Optional) A character value
| | |-avatar: (Optional) A character value
| | |-banned: A logical value
| | |-published: A character value
| | |-updated: (Optional) A character value
| | |-actor_id: A character value
| | |-bio: (Optional) A character value
| | |-local: A logical value
| | |-banner: (Optional) A character value
| | |-deleted: A logical value
| | |-inbox_url: A character value
| | |-matrix_user_id: (Optional) A character value
| | |-admin: A logical value
| | |-bot_account: A logical value
| | |-ban_expires: (Optional) A character value
| | `-instance_id: A numeric value
| |-community: <list>
| | |-id: A numeric value
| | |-name: A character value
| | |-title: A character value
| | |-description: (Optional) A character value
| | |-removed: A logical value
| | |-published: A character value
| | |-updated: (Optional) A character value
| | |-deleted: A logical value
| | |-nsfw: A logical value
| | |-actor_id: A character value
| | |-local: A logical value
| | |-icon: (Optional) A character value
| | |-banner: (Optional) A character value
| | |-followers_url: A character value
| | |-inbox_url: A character value
| | |-hidden: A logical value
| | |-posting_restricted_to_mods: A logical value
| | `-instance_id: A numeric value
| |-creator_banned_from_community: A logical value
| |-counts: <list>
| | |-id: A numeric value
| | |-post_id: A numeric value
| | |-comments: A numeric value
| | |-score: A numeric value
| | |-upvotes: A numeric value
| | |-downvotes: A numeric value
| | |-published: A character value
| | |-newest_comment_time_necro: A character value
| | |-newest_comment_time: A character value
| | |-featured_community: A logical value
| | |-featured_local: A logical value
| | |-hot_rank: A numeric value
| | `-hot_rank_active: A numeric value
| |-subscribed: One of "Subscribed", "NotSubscri...
| |-saved: A logical value
| |-read: A logical value
| |-creator_blocked: A logical value
| |-my_vote: (Optional) A numeric value
| `-unread_comments: A numeric value
|-community_view: <list>
| |-community: <list>
| | |-id: A numeric value
| | |-name: A character value
| | |-title: A character value
| | |-description: (Optional) A character value
| | |-removed: A logical value
| | |-published: A character value
| | |-updated: (Optional) A character value
| | |-deleted: A logical value
| | |-nsfw: A logical value
| | |-actor_id: A character value
| | |-local: A logical value
| | |-icon: (Optional) A character value
| | |-banner: (Optional) A character value
| | |-followers_url: A character value
| | |-inbox_url: A character value
| | |-hidden: A logical value
| | |-posting_restricted_to_mods: A logical value
| | `-instance_id: A numeric value
| |-subscribed: One of "Subscribed", "NotSubscri...
| |-blocked: A logical value
| `-counts: <list>
|   |-id: A numeric value
|   |-community_id: A numeric value
|   |-subscribers: A numeric value
|   |-posts: A numeric value
|   |-comments: A numeric value
|   |-published: A character value
|   |-users_active_day: A numeric value
|   |-users_active_week: A numeric value
|   |-users_active_month: A numeric value
|   |-users_active_half_year: A numeric value
|   `-hot_rank: A numeric value
|-moderators: <list>
| `-[Each element]: <list>
|   |-community: <list>
|   | |-id: A numeric value
|   | |-name: A character value
|   | |-title: A character value
|   | |-description: (Optional) A character value
|   | |-removed: A logical value
|   | |-published: A character value
|   | |-updated: (Optional) A character value
|   | |-deleted: A logical value
|   | |-nsfw: A logical value
|   | |-actor_id: A character value
|   | |-local: A logical value
|   | |-icon: (Optional) A character value
|   | |-banner: (Optional) A character value
|   | |-followers_url: A character value
|   | |-inbox_url: A character value
|   | |-hidden: A logical value
|   | |-posting_restricted_to_mods: A logical value
|   | `-instance_id: A numeric value
|   `-moderator: <list>
|     |-id: A numeric value
|     |-name: A character value
|     |-display_name: (Optional) A character value
|     |-avatar: (Optional) A character value
|     |-banned: A logical value
|     |-published: A character value
|     |-updated: (Optional) A character value
|     |-actor_id: A character value
|     |-bio: (Optional) A character value
|     |-local: A logical value
|     |-banner: (Optional) A character value
|     |-deleted: A logical value
|     |-inbox_url: A character value
|     |-matrix_user_id: (Optional) A character value
|     |-admin: A logical value
|     |-bot_account: A logical value
|     |-ban_expires: (Optional) A character value
|     `-instance_id: A numeric value
`-cross_posts: <list>
  `-[Each element]: <list>
    |-post: <list>
    | |-id: A numeric value
    | |-name: A character value
    | |-url: (Optional) A character value
    | |-body: (Optional) A character value
    | |-creator_id: A numeric value
    | |-community_id: A numeric value
    | |-removed: A logical value
    | |-locked: A logical value
    | |-published: A character value
    | |-updated: (Optional) A character value
    | |-deleted: A logical value
    | |-nsfw: A logical value
    | |-embed_title: (Optional) A character value
    | |-embed_description: (Optional) A character value
    | |-thumbnail_url: (Optional) A character value
    | |-ap_id: A character value
    | |-local: A logical value
    | |-embed_video_url: (Optional) A character value
    | |-language_id: A numeric value
    | |-featured_community: A logical value
    | `-featured_local: A logical value
    |-creator: <list>
    | |-id: A numeric value
    | |-name: A character value
    | |-display_name: (Optional) A character value
    | |-avatar: (Optional) A character value
    | |-banned: A logical value
    | |-published: A character value
    | |-updated: (Optional) A character value
    | |-actor_id: A character value
    | |-bio: (Optional) A character value
    | |-local: A logical value
    | |-banner: (Optional) A character value
    | |-deleted: A logical value
    | |-inbox_url: A character value
    | |-matrix_user_id: (Optional) A character value
    | |-admin: A logical value
    | |-bot_account: A logical value
    | |-ban_expires: (Optional) A character value
    | `-instance_id: A numeric value
    |-community: <list>
    | |-id: A numeric value
    | |-name: A character value
    | |-title: A character value
    | |-description: (Optional) A character value
    | |-removed: A logical value
    | |-published: A character value
    | |-updated: (Optional) A character value
    | |-deleted: A logical value
    | |-nsfw: A logical value
    | |-actor_id: A character value
    | |-local: A logical value
    | |-icon: (Optional) A character value
    | |-banner: (Optional) A character value
    | |-followers_url: A character value
    | |-inbox_url: A character value
    | |-hidden: A logical value
    | |-posting_restricted_to_mods: A logical value
    | `-instance_id: A numeric value
    |-creator_banned_from_community: A logical value
    |-counts: <list>
    | |-id: A numeric value
    | |-post_id: A numeric value
    | |-comments: A numeric value
    | |-score: A numeric value
    | |-upvotes: A numeric value
    | |-downvotes: A numeric value
    | |-published: A character value
    | |-newest_comment_time_necro: A character value
    | |-newest_comment_time: A character value
    | |-featured_community: A logical value
    | |-featured_local: A logical value
    | |-hot_rank: A numeric value
    | `-hot_rank_active: A numeric value
    |-subscribed: One of "Subscribed", "NotSubscri...
    |-saved: A logical value
    |-read: A logical value
    |-creator_blocked: A logical value
    |-my_vote: (Optional) A numeric value
    `-unread_comments: A numeric value
}
}
\description{
\code{HTTP.GET /post}
}
\examples{
\dontshow{if (remmy:::is_lemmy_world()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
post_data <- lemmy_get_post(1397803)

post_data$community_view$community$name

post_data$post_view$post$name
\dontshow{\}) # examplesIf}
}
