% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_get_report_count}
\alias{lemmy_get_report_count}
\title{Get counts for your reports}
\usage{
lemmy_get_report_count(
  community_id = NULL,
  auth,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{community_id}{(Optional) A numeric value.}

\item{auth}{A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 4:
\preformatted{<list>
|-community_id: (Optional) A numeric value
|-comment_reports: A numeric value
|-post_reports: A numeric value
`-private_message_reports: (Optional) A numeric value
}
}
\description{
\code{HTTP.GET /user/report_count}
}
