% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_get_unread_count}
\alias{lemmy_get_unread_count}
\title{Get your unread counts}
\usage{
lemmy_get_unread_count(
  auth,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{auth}{A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 3:

\if{html}{\out{<div class="sourceCode plaintext">}}\preformatted{<list>
├─replies: A numeric value
├─mentions: A numeric value
└─private_messages: A numeric value
}\if{html}{\out{</div>}}
}
\description{
\code{HTTP.GET /user/unread_count}
}
