% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_get_unread_registration_application_count}
\alias{lemmy_get_unread_registration_application_count}
\title{Get the unread registration applications count.}
\usage{
lemmy_get_unread_registration_application_count(
  auth,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{auth}{A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 1:
\preformatted{<list>
`-registration_applications: A numeric value
}
}
\description{
\code{HTTP.GET /admin/registration_application/count}
}
