% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_lock_post}
\alias{lemmy_lock_post}
\title{A moderator can lock a post ( IE disable new comments ).}
\usage{
lemmy_lock_post(
  post_id,
  locked,
  auth,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{post_id}{A numeric value.}

\item{locked}{A logical value.}

\item{auth}{A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 1:
\preformatted{<list>
`-post_view: <list>
  |-post: <list>
  | |-id: A numeric value
  | |-name: A character value
  | |-url: (Optional) A character value
  | |-body: (Optional) A character value
  | |-creator_id: A numeric value
  | |-community_id: A numeric value
  | |-removed: A logical value
  | |-locked: A logical value
  | |-published: A character value
  | |-updated: (Optional) A character value
  | |-deleted: A logical value
  | |-nsfw: A logical value
  | |-embed_title: (Optional) A character value
  | |-embed_description: (Optional) A character value
  | |-thumbnail_url: (Optional) A character value
  | |-ap_id: A character value
  | |-local: A logical value
  | |-embed_video_url: (Optional) A character value
  | |-language_id: A numeric value
  | |-featured_community: A logical value
  | `-featured_local: A logical value
  |-creator: <list>
  | |-id: A numeric value
  | |-name: A character value
  | |-display_name: (Optional) A character value
  | |-avatar: (Optional) A character value
  | |-banned: A logical value
  | |-published: A character value
  | |-updated: (Optional) A character value
  | |-actor_id: A character value
  | |-bio: (Optional) A character value
  | |-local: A logical value
  | |-banner: (Optional) A character value
  | |-deleted: A logical value
  | |-inbox_url: A character value
  | |-matrix_user_id: (Optional) A character value
  | |-admin: A logical value
  | |-bot_account: A logical value
  | |-ban_expires: (Optional) A character value
  | `-instance_id: A numeric value
  |-community: <list>
  | |-id: A numeric value
  | |-name: A character value
  | |-title: A character value
  | |-description: (Optional) A character value
  | |-removed: A logical value
  | |-published: A character value
  | |-updated: (Optional) A character value
  | |-deleted: A logical value
  | |-nsfw: A logical value
  | |-actor_id: A character value
  | |-local: A logical value
  | |-icon: (Optional) A character value
  | |-banner: (Optional) A character value
  | |-followers_url: A character value
  | |-inbox_url: A character value
  | |-hidden: A logical value
  | |-posting_restricted_to_mods: A logical value
  | `-instance_id: A numeric value
  |-creator_banned_from_community: A logical value
  |-counts: <list>
  | |-id: A numeric value
  | |-post_id: A numeric value
  | |-comments: A numeric value
  | |-score: A numeric value
  | |-upvotes: A numeric value
  | |-downvotes: A numeric value
  | |-published: A character value
  | |-newest_comment_time_necro: A character value
  | |-newest_comment_time: A character value
  | |-featured_community: A logical value
  | |-featured_local: A logical value
  | |-hot_rank: A numeric value
  | `-hot_rank_active: A numeric value
  |-subscribed: One of "Subscribed", "NotSubscri...
  |-saved: A logical value
  |-read: A logical value
  |-creator_blocked: A logical value
  |-my_vote: (Optional) A numeric value
  `-unread_comments: A numeric value
}
}
\description{
\code{HTTP.POST /post/lock}
}
