% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_login}
\alias{lemmy_login}
\title{Log into lemmy.}
\usage{
lemmy_login(
  username_or_email,
  password,
  totp_2fa_token = NULL,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{username_or_email}{A character value.}

\item{password}{A character value.}

\item{totp_2fa_token}{(Optional) A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 3:
\preformatted{<list>
|-jwt: (Optional) A character value
|-registration_created: A logical value
`-verify_email_sent: A logical value
}
}
\description{
\code{HTTP.POST /user/login}
}
