% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_mark_private_message_as_read}
\alias{lemmy_mark_private_message_as_read}
\title{Mark a private message as read.}
\usage{
lemmy_mark_private_message_as_read(
  private_message_id,
  read,
  auth,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{private_message_id}{A numeric value.}

\item{read}{A logical value.}

\item{auth}{A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 1:
\preformatted{<list>
`-private_message_view: <list>
  |-private_message: <list>
  | |-id: A numeric value
  | |-creator_id: A numeric value
  | |-recipient_id: A numeric value
  | |-content: A character value
  | |-deleted: A logical value
  | |-read: A logical value
  | |-published: A character value
  | |-updated: (Optional) A character value
  | |-ap_id: A character value
  | `-local: A logical value
  |-creator: <list>
  | |-id: A numeric value
  | |-name: A character value
  | |-display_name: (Optional) A character value
  | |-avatar: (Optional) A character value
  | |-banned: A logical value
  | |-published: A character value
  | |-updated: (Optional) A character value
  | |-actor_id: A character value
  | |-bio: (Optional) A character value
  | |-local: A logical value
  | |-banner: (Optional) A character value
  | |-deleted: A logical value
  | |-inbox_url: A character value
  | |-matrix_user_id: (Optional) A character value
  | |-admin: A logical value
  | |-bot_account: A logical value
  | |-ban_expires: (Optional) A character value
  | `-instance_id: A numeric value
  `-recipient: <list>
    |-id: A numeric value
    |-name: A character value
    |-display_name: (Optional) A character value
    |-avatar: (Optional) A character value
    |-banned: A logical value
    |-published: A character value
    |-updated: (Optional) A character value
    |-actor_id: A character value
    |-bio: (Optional) A character value
    |-local: A logical value
    |-banner: (Optional) A character value
    |-deleted: A logical value
    |-inbox_url: A character value
    |-matrix_user_id: (Optional) A character value
    |-admin: A logical value
    |-bot_account: A logical value
    |-ban_expires: (Optional) A character value
    `-instance_id: A numeric value
}
}
\description{
\code{HTTP.POST /private_message/mark_as_read}
}
