% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_password_reset}
\alias{lemmy_password_reset}
\title{Reset your password.}
\usage{
lemmy_password_reset(
  email,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{email}{A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 0:
\preformatted{<list>
}
}
\description{
\code{HTTP.POST /user/password_reset}
}
