% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_purge_community}
\alias{lemmy_purge_community}
\title{Purge / Delete a community from the database.}
\usage{
lemmy_purge_community(
  community_id,
  reason = NULL,
  auth,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{community_id}{A numeric value.}

\item{reason}{(Optional) A character value.}

\item{auth}{A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 1:
\preformatted{<list>
`-success: A logical value
}
}
\description{
\code{HTTP.POST /admin/purge/community}
}
