% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_register}
\alias{lemmy_register}
\title{Register a new user.}
\usage{
lemmy_register(
  username,
  password,
  password_verify,
  show_nsfw,
  email = NULL,
  captcha_uuid = NULL,
  captcha_answer = NULL,
  honeypot = NULL,
  answer = NULL,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{username}{A character value.}

\item{password}{A character value.}

\item{password_verify}{A character value.}

\item{show_nsfw}{A logical value.}

\item{email}{(Optional) A character value.}

\item{captcha_uuid}{(Optional) A character value.}

\item{captcha_answer}{(Optional) A character value.}

\item{honeypot}{(Optional) A character value.}

\item{answer}{(Optional) A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 3:
\preformatted{<list>
|-jwt: (Optional) A character value
|-registration_created: A logical value
`-verify_email_sent: A logical value
}
}
\description{
\code{HTTP.POST /user/register}
}
