% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_save_user_settings}
\alias{lemmy_save_user_settings}
\title{Save your user settings.}
\usage{
lemmy_save_user_settings(
  show_nsfw = NULL,
  show_scores = NULL,
  theme = NULL,
  default_sort_type = NULL,
  default_listing_type = NULL,
  interface_language = NULL,
  avatar = NULL,
  banner = NULL,
  display_name = NULL,
  email = NULL,
  bio = NULL,
  matrix_user_id = NULL,
  show_avatars = NULL,
  send_notifications_to_email = NULL,
  bot_account = NULL,
  show_bot_accounts = NULL,
  show_read_posts = NULL,
  show_new_post_notifs = NULL,
  discussion_languages = NULL,
  generate_totp_2fa = NULL,
  auth,
  open_links_in_new_tab = NULL,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{show_nsfw}{(Optional) A logical value.}

\item{show_scores}{(Optional) A logical value.}

\item{theme}{(Optional) A character value.}

\item{default_sort_type}{(Optional) One of "Active", "Hot", "New", "Old", "TopDay", "TopWeek", "TopMonth", "TopYear", "TopAll", "MostComments", "NewComments", "TopHour", "TopSixHour", "TopTwelveHour", "TopThreeMonths", "TopSixMonths", or "TopNineMonths".}

\item{default_listing_type}{(Optional) One of "All", "Local", or "Subscribed".}

\item{interface_language}{(Optional) A character value.}

\item{avatar}{(Optional) A character value.}

\item{banner}{(Optional) A character value.}

\item{display_name}{(Optional) A character value.}

\item{email}{(Optional) A character value.}

\item{bio}{(Optional) A character value.}

\item{matrix_user_id}{(Optional) A character value.}

\item{show_avatars}{(Optional) A logical value.}

\item{send_notifications_to_email}{(Optional) A logical value.}

\item{bot_account}{(Optional) A logical value.}

\item{show_bot_accounts}{(Optional) A logical value.}

\item{show_read_posts}{(Optional) A logical value.}

\item{show_new_post_notifs}{(Optional) A logical value.}

\item{discussion_languages}{(Optional) A numeric vector.}

\item{generate_totp_2fa}{(Optional) A logical value.}

\item{auth}{A character value.}

\item{open_links_in_new_tab}{(Optional) A logical value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 3:
\preformatted{<list>
|-jwt: (Optional) A character value
|-registration_created: A logical value
`-verify_email_sent: A logical value
}
}
\description{
\code{HTTP.PUT /user/save_user_settings}
}
