% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_verify_email}
\alias{lemmy_verify_email}
\title{Verify your email}
\usage{
lemmy_verify_email(
  token,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{token}{A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 0:
\preformatted{<list>
}
}
\description{
\code{HTTP.POST /user/verify_email}
}
