% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-methods.R
\name{lemmy_get_captcha}
\alias{lemmy_get_captcha}
\title{Fetch a Captcha.}
\usage{
lemmy_get_captcha(
  auth = NULL,
  .lemmy_instance = getOption("lemmy_instance", "https://lemmy.world")
)
}
\arguments{
\item{auth}{(Optional) A character value.}

\item{.lemmy_instance}{Default: "https://lemmy.world". Change the default instance by setting the \code{lemmy_instance} \code{\link[=options]{options()}}.}
}
\value{
A \code{list} of length 1:
\preformatted{<list>
`-ok: <list>
  |-png: A character value
  |-wav: A character value
  `-uuid: A character value
}
}
\description{
\code{HTTP.GET /user/get_captcha}
}
